Remove outdated library and its modules.
