+++ Backend/app/loan_management/book_loans_router.py
@@ -1,12 +1,14 @@
 from typing import Any
 from fastapi import APIRouter, Depends, HTTPException
-from app.library.library_models import PhysicalCopyData
-from app.library.library_service import BookNotFound, NoCopiesAvailable
-from app.loan_management.book_loans_service import LoanService
+from app.loan_management.book_loans_service import (
+    BookNotFound,
+    LoanService,
+    NoCopiesAvailable,
+)
 from app.loan_management.book_loans_dtos import (
     LoanDTO,
     LoanInformationDTO,
-    RequestedBookDTO,
+    PhysicalCopyDTO,
 )
 from fastapi.security import HTTPBearer
 from ..middlewares import verify_token
@@ -23,7 +25,7 @@ loan_service = LoanService(DATABASE_PATH, CATALOGUE_PATH)
 licence_service = LicenceService(DATABASE_PATH, CATALOGUE_PATH)
 
 
-@router.post("/borrow", response_model=PhysicalCopyData)
+@router.post("/borrow", response_model=PhysicalCopyDTO)
 async def create_requested_book(book: LoanDTO, token=Depends(HTTPBearer())):
     user_data: TokenDataDTO = await verify_token(token.credentials)
     requested_book = licence_service.consult_book_data(book.isbn)
