+++ Backend/app/loan_management/books_router.py
