+++ Frontend/src/components/ShowLoans.tsx
@@ -17,7 +17,6 @@ const ShowLoans = () => {
 
         const data = JSON.parse(jsonData);
         setBooks(data);
-        console.log(data);
       } catch (error) {
         console.error("Error al cargar los datos desde AsyncStorage:", error);
       }
