+++ Frontend/src/services/LibraryService.ts
@@ -14,8 +14,6 @@ export class LibraryService {
         method: "GET",
       });
 
-      console.log(books)
-
       return books;
     } catch (error) {
       console.error("Error fetching books:", error);
